proj/hps_accel: move gateware to gen1 subdirectory
Moves gateware to a sub-directory gateware/gen1 in anticipation of there
being a gen2 sub-directory in future.
