+++ proj/hps_accel/gateware/gen1/README.md
@@ -1,0 +1,3 @@
+# gen1 accelerator
+
+This directory contains accelerator gateware for the gen1 accelerator.
