+++ proj/hps_accel/gateware/gen1/test_mem.py
