+++ proj/hps_accel/gateware/gen1/test_sp_mem.py
